+++ includes/bsq_time.h
@@ -1,8 +1,0 @@
-#ifndef BSQ_TIME_H
-# define BSQ_TIME_H
-
-# include "bsq.h"
-
-void	print_exec_time(t_exec_time *exec_time);
-
-#endif